working on ingredients/instructions

+++ allSpiceSharp.client/src/components/RecipeDetailsModal.vue
@@ -6,31 +6,38 @@
           <img :src="activeRecipe.img" alt="" class="img-fluid selectable recipe-img rounded">
         </div>
         <div class="col-8">
-          <div class="row justify-content-end">
-            <div class="col-1">
-              <button @click="favoriteRecipe" class="btn text-success fs-5" title="Favorite Recipe"><i
-                  class="mdi mdi-heart"></i></button>
+          <div class="row">
+            <div class="col-9">
+              <h2 class="text-success">{{ activeRecipe.title }}</h2>
+              <h5 class="text-primary">{{ activeRecipe.category }}</h5>
             </div>
             <div class="col-1">
-              <button @click="deleteRecipe" class="btn text-danger fs-5" title="Delete Recipe"><i
-                  class="mdi mdi-delete"></i></button>
+              <button v-if="!recipe.favoriteId" @click="favoriteRecipe" class="btn text-success fs-5"
+                title="Favorite Recipe"><i class="mdi mdi-heart"></i></button>
+              <button v-else class="btn text-success fs-5" title="Unfavorite Recipe"><i
+                  class="mdi mdi-heart-broken"></i></button>
+            </div>
+            <div class="col-1">
+              <button v-if="account.id == recipe.creatorId" @click="deleteRecipe" class="btn text-danger fs-5"
+                title="Delete Recipe"><i class="mdi mdi-delete"></i></button>
             </div>
             <button type="button" class="btn-close me-2 mt-2" data-bs-dismiss="modal" aria-label="Close"></button>
           </div>
-          <div class="row">
-            <h2 class="text-success">{{ activeRecipe.title }}</h2>
-            <h4>{{ activeRecipe.category }}</h4>
-          </div>
-          <div class="row">
-            <div class="col-6 card p-0">
+
+          <div class="row justify-content-evenly">
+            <div class="col-5 card p-2 m-1">
               <h5 class="bg-success text-center py-1">Recipe Ingredients</h5>
               <div v-for="i in ingredients" :key="i.id">
                 <Ingredients :ingredient="i" />
               </div>
             </div>
-            <div class="col-6 card p-0">
+            <div class="col-5 card p-2 m-1">
               <h5 class="bg-success text-center py-1">Recipe Instructions</h5>
               <div class="">{{ activeRecipe.instructions }}</div>
+              <form>
+                <input placeholder="Add Next Step" />
+                <button @click="" class="text-success" title="add instruction"><i class="mdi mdi-plus"></i></button>
+              </form>
             </div>
           </div>
           <div class="row">
@@ -66,9 +73,13 @@ export default {
       async deleteRecipe() {
         try {
           // logger.log('click delete', AppState.activeRecipe.id)
-          const recipeId = AppState.activeRecipe.id
-          await recipesService.deleteRecipe(recipeId)
-          Modal.getOrCreateInstance('#recipeDetailModal').hide()
+          if (await Pop.confirm('Are you sure you want to delete this recipe?')) {
+            const recipeId = AppState.activeRecipe.id
+            await recipesService.deleteRecipe(recipeId)
+            Modal.getOrCreateInstance('#recipeDetailModal').hide()
+            Pop.success('Recipe has been deleted.')
+          }
+
         } catch (error) {
           logger.error(error)
           Pop.error(error.message)
@@ -77,9 +88,10 @@ export default {
 
       async favoriteRecipe() {
         try {
-          logger.log('clicked heart', AppState.activeRecipe.id)
-          const favoriteData = { recipeId: AppState.activeRecipe.id }
-          await favoritesService.favoriteRecipe(favoriteData)
+          // logger.log('clicked heart', AppState.activeRecipe.id)
+          const recipeId = { recipeId: AppState.activeRecipe.id }
+          await favoritesService.favoriteRecipe(recipeId)
+          Pop.success('Recipe is now in your favorites!')
         } catch (error) {
           logger.error(error)
           Pop.error(error.message)
